requirements_common_team_train.txt
numpy
pandas
scikit-learn
pillow
tqdm
timm
albumentations
opencv-python-headless
